+++ docs/releases.md
@@ -4,6 +4,16 @@ uid: releases
 
 # Releases
 
+## [4.5.2](https://github.com/BEagle1984/silverback/releases/tag/v4.6.0)
+
+### What's new
+
+* Add `GetOffsetsForTimestamp` to `KafkaConsumer`
+
+### Fixes
+
+* Fix retry policy when batch processing all partitions together
+
 ## [4.5.1](https://github.com/BEagle1984/silverback/releases/tag/v4.5.1)
 
 ### Fixes